call SLAS to get token and authenticate
